requirements.txt
Pillow==5.2.0
selenium==3.14.0
urllib3==1.23
